Edit for Categories and Chapters + small changes in design

+++ src/main/java/com/unibuc/mds/memoreasy/Controllers/DeleteCategoryController.java
@@ -1,5 +1,4 @@
 package com.unibuc.mds.memoreasy.Controllers;
-
 import com.unibuc.mds.memoreasy.Utils.DatabaseUtils;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -9,27 +8,31 @@ import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.stage.Stage;
 import org.kordamp.bootstrapfx.BootstrapFX;
-
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
-
 import static javafx.fxml.FXMLLoader.load;
 
 public class DeleteCategoryController {
-    static int idCategory=0;
+    private int idCategory;
 
+    @FXML
+    private Button deleteButton;
+
+    //Ma intorc la categoria corespunzatoare pe care am primit-o prin acea pereche (nume, id).
     public void delete(ActionEvent event) throws SQLException, IOException {
-        String sql1 = "delete from flashcard where id_chapter in (select id_chapter from  chapter where id_category = "+ idCategory+")";
-        String sql2 = "delete from chapter where id_category = "+ idCategory;
-        String sql3 = "delete from category where id_category = "+ idCategory;
-        Connection con = DatabaseUtils.getConnection();
-        Statement stmt = con.createStatement();
-        stmt.executeUpdate(sql1);
-        stmt.executeUpdate(sql2);
-        stmt.executeUpdate(sql3);
-        con.close();
+        if (event.getSource() == deleteButton) {
+            String sql1 = "delete from flashcard where id_chapter in (select id_chapter from  chapter where id_category = " + idCategory + ")";
+            String sql2 = "delete from chapter where id_category = " + idCategory;
+            String sql3 = "delete from category where id_category = " + idCategory;
+            Connection con = DatabaseUtils.getConnection();
+            Statement stmt = con.createStatement();
+            stmt.executeUpdate(sql1);
+            stmt.executeUpdate(sql2);
+            stmt.executeUpdate(sql3);
+            con.close();
+        }
 
         Parent root = load(getClass().getResource("/com/unibuc/mds/memoreasy/Views/AllCategories/AllCategoriesView.fxml"));
         Stage stage = (Stage) ((Node)event.getSource()).getScene().getWindow();
@@ -39,7 +42,7 @@ public class DeleteCategoryController {
         stage.show();
     }
 
-    static public void setIdCategory(int id_category) {
+    public void setIdCategory(int id_category) {
         idCategory = id_category;
     }
 }
